[ 2.2.0 ] Class structure normalized ...

+++ merge-HAR/Data.cpp
@@ -1,12 +1,12 @@
 #include "Data.h"
 #include "MANode.h"
 
-long int CData::ID_COUNT = 0;
-long int CData::ARRIVAL_COUNT = 0;
-long double CData::DELAY_SUM = 0;
-long int CData::OVERFLOW_COUNT = 0;
-long int CData::DELIVERY_AT_HOTSPOT_COUNT = 0;
-long int CData::DELIVERY_ON_ROUTE_COUNT = 0;
+int CData::ID_COUNT = 0;
+int CData::ARRIVAL_COUNT = 0;
+double CData::DELAY_SUM = 0;
+int CData::OVERFLOW_COUNT = 0;
+int CData::DELIVERY_AT_HOTSPOT_COUNT = 0;
+int CData::DELIVERY_ON_ROUTE_COUNT = 0;
 
 CData::~CData(void)
 {
