+++ merge-HAR/Position.cpp
@@ -1,6 +1,6 @@
 #include "Position.h"
 
-long int CPosition::ID_COUNT = 0;
+int CPosition::ID_COUNT = 0;
 int CPosition::nPositions = 0;
 vector<CPosition *> CPosition::positions;
 vector<CPosition *> CPosition::deletedPositions;
